Create helper files
