+++ src/Conversion/ONNXToKrnl/ONNXToKrnlCommon.cpp
@@ -11,36 +11,6 @@
 
 #include "src/Conversion/ONNXToKrnl/ONNXToKrnlCommon.hpp"
 
-// Identity affine
-AffineMap getIdentityDimMap(Builder &builder) {
-  return AffineMap::get(1, 0, {builder.getAffineDimExpr(0)});
-}
-
-// Pool/conv affine
-// dim =
-//   let numerator = (input + pad - (kernel - 1) * dilation - 1)
-//   in let denominator = stride
-//      in
-//        if (ceilMode)
-//          ceil(numerator / denominator) + 1
-//        else
-//          floor(numerator / denominator) + 1
-AffineMap getConvDimMap(Builder &builder, bool ceilMode) {
-  AffineExpr input = builder.getAffineDimExpr(0);
-  AffineExpr kernel = builder.getAffineSymbolExpr(0);
-  AffineExpr pad = builder.getAffineSymbolExpr(1);
-  AffineExpr stride = builder.getAffineSymbolExpr(2);
-  AffineExpr dilation = builder.getAffineSymbolExpr(3);
-
-  AffineExpr dimExp;
-  if (ceilMode)
-    dimExp = (input + pad - (kernel - 1) * dilation - 1).ceilDiv(stride) + 1;
-  else
-    dimExp = (input + pad - (kernel - 1) * dilation - 1).floorDiv(stride) + 1;
-
-  return AffineMap::get(1, 4, {dimExp});
-}
-
 /// Check is all dimensions are known at compile time.
 bool hasAllConstantDimensions(MemRefType type) {
   auto memRefShape = type.getShape();
